+++ randomizer-core/src/main/java/de/alpharogroup/random/RandomObjectsExtensions.java
@@ -24,10 +24,15 @@
  */
 package de.alpharogroup.random;
 
+import java.lang.reflect.Field;
+import java.lang.reflect.Modifier;
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import de.alpharogroup.reflection.ReflectionExtensions;
+import lombok.NonNull;
 import lombok.experimental.UtilityClass;
 
 /**
@@ -251,6 +256,113 @@ public final class RandomObjectsExtensions
 			}
 		}
 		return sb.toString();
+	}
+
+	/**
+	 * Factory method for create a new random object of the given {@link Class}
+	 *
+	 * @param <T>
+	 *            the generic type
+	 * @param cls
+	 *            the class
+	 * @return the new random object
+	 * 
+	 * @throws IllegalAccessException
+	 *             is thrown if the class or its default constructor is not accessible.
+	 * @throws InstantiationException
+	 *             is thrown if this {@code Class} represents an abstract class, an interface, an
+	 *             array class, a primitive type, or void; or if the class has no default
+	 *             constructor; or if the instantiation fails for some other reason.
+	 * @throws NoSuchFieldException
+	 *             is thrown if no such field exists
+	 */
+	public static <T> T newRandomObject(final @NonNull Class<T> cls)
+		throws IllegalAccessException, InstantiationException, NoSuchFieldException
+	{
+		T instance = ReflectionExtensions.newInstance(cls);
+		Field[] allDeclaredFields = ReflectionExtensions.getAllDeclaredFields(cls);
+		for (Field field : allDeclaredFields)
+		{
+			if (Modifier.isFinal(field.getModifiers()))
+			{
+				continue;
+			}
+			Object value = newRandomValue(field);
+			ReflectionExtensions.setFieldValue(instance, field.getName(), value);
+		}
+		return instance;
+	}
+
+	/**
+	 * Factory method for create a new random value for the given {@link Field field}
+	 *
+	 * @param field
+	 *            the field
+	 * @return the new random value
+	 * @throws IllegalAccessException
+	 *             is thrown if the class or its default constructor is not accessible.
+	 * @throws InstantiationException
+	 *             is thrown if this {@code Class} represents an abstract class, an interface, an
+	 *             array class, a primitive type, or void; or if the class has no default
+	 *             constructor; or if the instantiation fails for some other reason.
+	 * @throws NoSuchFieldException
+	 *             is thrown if no such field exists
+	 */
+	public static Object newRandomValue(Field field)
+		throws IllegalAccessException, InstantiationException, NoSuchFieldException
+	{
+		Class<?> type = field.getType();
+		if (type.isEnum())
+		{
+			Enum<?> randomEnum = RandomExtensions
+				.getRandomEnumFromClassname(type.getCanonicalName());
+			return randomEnum;
+		}
+		else if (type.equals(Void.TYPE) || type.equals(Void.class))
+		{
+			return null;
+		}
+		else if (type.equals(Byte.TYPE) || type.equals(Byte.class))
+		{
+			return Byte.valueOf(RandomExtensions.randomByte());
+		}
+		else if (type.equals(Character.TYPE) || type.equals(Character.class))
+		{
+			return Character.valueOf(RandomExtensions.randomChar());
+		}
+		else if (type.equals(Short.TYPE) || type.equals(Short.class))
+		{
+			return Short.valueOf(RandomExtensions.randomShort());
+		}
+		else if (type.equals(Boolean.TYPE) || type.equals(Boolean.class))
+		{
+			return Boolean.valueOf(RandomExtensions.randomBoolean());
+		}
+		else if (type.equals(Integer.TYPE) || type.equals(Integer.class))
+		{
+			return RandomExtensions.randomInt();
+		}
+		else if (type.equals(Long.TYPE) || type.equals(Long.class))
+		{
+			return RandomExtensions.randomLong();
+		}
+		else if (type.equals(Double.TYPE) || type.equals(Double.class))
+		{
+			return RandomExtensions.randomDouble();
+		}
+		else if (type.equals(Float.TYPE) || type.equals(Float.class))
+		{
+			return RandomExtensions.randomFloat();
+		}
+		else if (type.equals(String.class))
+		{
+			return RandomExtensions.randomUUID().toString();
+		}
+		else if (type.equals(BigInteger.class))
+		{
+			return RandomExtensions.randomSerialNumber();
+		}
+		return newRandomObject(type);
 	}
 
 }