+++ randomizer-core/src/test/java/de/alpharogroup/random/NullSecureRandomExtensionsTest.java
@@ -171,13 +171,13 @@ public class NullSecureRandomExtensionsTest extends BaseTestCase
 	}
 
 	/**
-	 * Test method for {@link RandomExtensions#getRandomEnum(Enum)} .
+	 * Test method for {@link RandomExtensions#getRandomEnumFromObject(Enum)} .
 	 */
 	@Test
 	public void testGetRandomEnum()
 	{
 		final Gender enumEntry = Gender.FEMALE;
-		final Gender randomEnumEntry = RandomExtensions.getRandomEnum(enumEntry);
+		final Gender randomEnumEntry = RandomExtensions.getRandomEnumFromObject(enumEntry);
 
 		final Gender[] genders = Gender.values();
 		assertTrue("Enum value should contain the random value.",
@@ -185,24 +185,24 @@ public class NullSecureRandomExtensionsTest extends BaseTestCase
 	}
 
 	/**
-	 * Test method for {@link RandomExtensions#getRandomEnum(Enum[])} .
+	 * Test method for {@link RandomExtensions#getRandomEnumFromEnumValues(Enum[])} .
 	 */
 	@Test
 	public void testGetRandomEnumArray()
 	{
 		final Gender[] genders = Gender.values();
-		final Gender randomEnumEntry = RandomExtensions.getRandomEnum(genders);
+		final Gender randomEnumEntry = RandomExtensions.getRandomEnumFromEnumValues(genders);
 		assertTrue("Enum value should contain the random value.",
 			ArrayUtils.contains(genders, randomEnumEntry));
 	}
 
 	/**
-	 * Test method for {@link RandomExtensions#getRandomEnum(Enum)} .
+	 * Test method for {@link RandomExtensions#getRandomEnumFromObject(Enum)} .
 	 */
 	@Test
 	public void testGetRandomEnumClass()
 	{
-		final Gender randomEnumEntry = RandomExtensions.getRandomEnum(Gender.class);
+		final Gender randomEnumEntry = RandomExtensions.getRandomEnumFromClass(Gender.class);
 
 		final Gender[] genders = Gender.values();
 		assertTrue("Enum value should contain the random value.",
@@ -210,13 +210,13 @@ public class NullSecureRandomExtensionsTest extends BaseTestCase
 	}
 
 	/**
-	 * Test method for {@link RandomExtensions#getRandomEnum(String)} .
+	 * Test method for {@link RandomExtensions#getRandomEnumFromClassname(String)} .
 	 */
 	@Test
 	public void testGetRandomEnumString()
 	{
 		final String enumClassName = "de.alpharogroup.test.objects.enums.Gender";
-		final Gender randomEnumEntry = RandomExtensions.getRandomEnum(enumClassName);
+		final Gender randomEnumEntry = RandomExtensions.getRandomEnumFromClassname(enumClassName);
 
 		final Gender[] genders = Gender.values();
 		assertTrue("Enum value should contain the random value.",